requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
colorama>=0.4.6
